modified edge case avatar

+++ controller/users_controller.js
@@ -33,7 +33,11 @@ module.exports.update= async function(req,res){
 
                 if(req.file){
                     if(user.avatar){
-                        fs.unlinkSync(path.join(__dirname,'..',user.avatar));   
+                        //if there are no previously updated avatar it will not show error upon adding new avatar
+                        if (fs.existsSync(path.join(__dirname,'..',user.avatar))) {
+                            fs.unlinkSync(path.join(__dirname,'..',user.avatar));   
+                        }
+                       
                     }
                     //this is saving the path of uploaded file into the avatar field in the user 
                     user.avatar=User.avatarPath+'/'+req.file.filename;
